Fixes for new anal g++ in Ubuntu 10.04

+++ src/queryablehost.cpp
@@ -29,7 +29,7 @@ void QueryableHost::initialize_result_set(map<string, ResultSet> & rs, QueryRow&
 
 bool QueryableHost::query_snmp(SNMP& snmp_session, QueryRow& row, map<string, ResultSet>& rs)
 {
-        uint64_t value;
+        unsigned long long value;
         time_t response_time;
         bool success = snmp_session.get_counter(row.oid, &value, &response_time);
         if (success) {
@@ -87,7 +87,7 @@ map<string, ResultSet> QueryableHost::get_all_resultsets()
 // Calculate the traffic rate between to points, for a given counter size (bits).
 // bits == 0 means it's a 32 bit gauge (RTG legacy).
 // bits == 32 or 64 means it's that size of counter.
-pair<uint64_t, uint64_t> QueryableHost::calculate_rate(time_t prev_time, uint64_t prev_counter, time_t cur_time, uint64_t cur_counter, int bits)
+pair<unsigned long long, unsigned long long> QueryableHost::calculate_rate(time_t prev_time, unsigned long long prev_counter, time_t cur_time, unsigned long long cur_counter, int bits)
 {
         time_t time_diff = cur_time - prev_time;
         if (time_diff == 0) {
@@ -99,7 +99,7 @@ pair<uint64_t, uint64_t> QueryableHost::calculate_rate(time_t prev_time, uint64_
                      << "bits " << bits << endl;
                 exit(-1);
         }
-        uint64_t counter_diff = cur_counter - prev_counter;
+        unsigned long long counter_diff = cur_counter - prev_counter;
         if (prev_counter > cur_counter) {
                 // We seem to have a wrap.
                 // Wrap it back to find the correct rate.
@@ -111,10 +111,10 @@ pair<uint64_t, uint64_t> QueryableHost::calculate_rate(time_t prev_time, uint64_
 
         if (bits == 0)
                 // It's a gauge so just return the value as both counter diff and rate.
-                return pair<uint64_t, uint64_t> (cur_counter, cur_counter);
+                return pair<unsigned long long, unsigned long long> (cur_counter, cur_counter);
         else
                 // Return the calculated rate.
-                return pair<uint64_t, uint64_t> (counter_diff, counter_diff / time_diff);
+                return pair<unsigned long long, unsigned long long> (counter_diff, counter_diff / time_diff);
 }
 
 // Query all targets for a host, process rates compared with cache, and return vector of database queries
@@ -179,10 +179,10 @@ string QueryableHost::build_insert_query(ResultSet& r)
                 if (cache.times.find(key) != cache.times.end()) {
                         // We have a cache entry, so we can calculate rate since last measurement.
                         time_t prev_time = cache.times[key];
-                        uint64_t prev_counter = cache.counters[key];
+                        unsigned long long prev_counter = cache.counters[key];
 
                         // Get the rate, corrected for wraps etc.
-                        pair<uint64_t, uint64_t> rate = calculate_rate(prev_time, prev_counter, row.dtime, row.counter, row.bits);
+                        pair<unsigned long long, unsigned long long> rate = calculate_rate(prev_time, prev_counter, row.dtime, row.counter, row.bits);
 
                         // Verify that the resulting value is reasonable, i.e. lower than interface speed.
                         if (rate.second <= row.speed) {
